Image Processing Final Project
In this project, I applied fundamental image processing techniques using Python and OpenCV to solve a series of real-world image-related problems. Each task focused on a different concept, such as pixel manipulation, image enhancement, noise reduction, and transformation. The goal was to analyze, modify, and interpret various grayscale and color images from the provided dataset.

In this project, I tried to obtain different images from the images we see with our eyes in the world using a set of image processing techniques.
Our goal is to analyze and interpret the grayscale and color images with the data we have.